Rename some functions
