+++ e2e_tests/visual_regression_test.py
@@ -2,7 +2,7 @@ import optuna
 from playwright.sync_api import Page
 import pytest
 
-from .mock_db import make_test_server
+from .test_server import make_test_server
 
 
 study_names = [
@@ -21,7 +21,7 @@ study_names = [
 ]
 
 
-def make_dummy_storage(study_name: str) -> optuna.storages.InMemoryStorage:
+def make_test_storage(study_name: str) -> optuna.storages.InMemoryStorage:
     storage = optuna.storages.InMemoryStorage()
     sampler = optuna.samplers.RandomSampler(seed=0)
 
@@ -213,7 +213,7 @@ def make_dummy_storage(study_name: str) -> optuna.storages.InMemoryStorage:
 @pytest.fixture(params=study_names)
 def storage(request: pytest.FixtureRequest) -> optuna.storages.InMemoryStorage:
     study_name = request.param
-    storage = make_dummy_storage(study_name)
+    storage = make_test_storage(study_name)
     return storage
